+++ src/index.ts
@@ -16,7 +16,7 @@ type Position = [number, number, number] | [number, number];
  * The value of the bbox member MUST be an array of
  * length 2*n where n is the number of dimensions represented in the
  * contained geometries, with all axes of the most southwesterly point
- * followed by all axes of the more northeasterly point.  The axes order
+ * followed by all axes of the more northeasterly point. The axes order
  * of a bbox follows the axes order of geometries.
  */
 type BoundingBox = [number, number, number, number] | [number, number, number, number, number, number];
@@ -178,6 +178,48 @@ export class Point extends Geometry {
     return this.elevation === null || isNaN(this.elevation)
       ? [this.longitude, this.latitude]
       : [this.longitude, this.latitude, this.elevation];
+  }
+
+  /**
+   *
+   * @param {Point} otherPoint to get bearing to
+   * @returns {number} in Nautical miles
+   */
+  getVectorTo(otherPoint: Point): Vector {
+    const lat1 = (this.latitude / 180) * Math.PI;
+    const lon1 = (this.longitude / 180) * Math.PI;
+    const lat2 = (otherPoint.latitude / 180) * Math.PI;
+    const lon2 = (otherPoint.longitude / 180) * Math.PI;
+
+    const dLon = lon2 - lon1;
+    const dLat = lat2 - lat1;
+
+    const y = Math.sin(dLon) * Math.cos(lat2);
+    const x = Math.cos(lat1) * Math.sin(lat2) - Math.sin(lat1) * Math.cos(lat2) * Math.cos(dLon);
+    const bearing = ((Math.atan2(y, x) * 180) / Math.PI + 360) % 360;
+
+    const a =
+      Math.sin(dLat / 2) * Math.sin(dLat / 2) +
+      Math.sin(dLon / 2) * Math.sin(dLon / 2) * Math.cos(lat1) * Math.cos(lat2);
+    const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
+    const meters = 6_371_000 * c;
+
+    return new Vector(meters, bearing);
+  }
+
+  getPointBy(vector: Vector): Point {
+    const d = vector.meters;
+    const brng = (((vector.bearing + 360) % 360) / 180) * Math.PI;
+    const lat1 = (this.latitude / 180) * Math.PI;
+    const lon1 = (this.longitude / 180) * Math.PI;
+    const R = 6_371_000;
+
+    const lat2 = Math.asin(Math.sin(lat1) * Math.cos(d / R) + Math.cos(lat1) * Math.sin(d / R) * Math.cos(brng));
+    const lon2 =
+      lon1 +
+      Math.atan2(Math.sin(brng) * Math.sin(d / R) * Math.cos(lat1), Math.cos(d / R) - Math.sin(lat1) * Math.sin(lat2));
+
+    return new Point((lon2 * 180) / Math.PI, (lat2 * 180) / Math.PI, this.elevation);
   }
 }
 
@@ -408,6 +450,30 @@ export class FeatureCollection extends GeoJSON {
   }
 }
 
+export class Vector {
+  meters: number;
+  protected _bearing!: number;
+
+  constructor(meters: number, bearing: number) {
+    this.meters = meters;
+    this.bearing = bearing;
+  }
+
+  /**
+   * @returns {number} 0..360
+   */
+  get bearing(): number {
+    return this._bearing;
+  }
+
+  /**
+   * @param {number} bearing 0..360
+   */
+  set bearing(bearing: number) {
+    this._bearing = bearing % 360;
+  }
+}
+
 export default {
   Point,
   MultiPoint,
@@ -418,4 +484,5 @@ export default {
   GeometryCollection,
   Feature,
   FeatureCollection,
+  Vector,
 };